ci: deploy Go service to DigitalOcean App Platform
Sets up a Terraform pipeline to provision a DigitalOcean App Platform
app. Terraform will execute a plan on PR branches but will only apply
changes when the PR is merged. Once Terraform has succeeded, a job is
triggered to deploy the latest state of the Go service to DigitalOcean.

A couple of services were provisioned as part of this change:

- A Terraform Cloud account was created (under my name, will be
  transferred once the foundation is set up) to run the pipeline and
  host the .tfstate file. A TF_API_TOKEN was generated and added to the
  repo secrets.

- A DigitalOcean account was created (also under my name and will be
  eventually transferred). A DIGITALOCEAN_ACCESS_TOKEN was generated and
  added to the repo secrets.