ci: use api.omshub.org as the domain for the DO app
Adds a domain configuration to the DO app which enables us to add a
CNAME record in Google Domains to route api.omshub.org to the Go
service.

Currently, the CNAME record needs to be manually set.
https://github.com/omshub/core-api/issues/16#issuecomment-1129642799
outlines some advantages of Cloudflare which we can opt to use later
down the track.