fix: install deps on startup to fix perm issue
Moves the commands to install deps like delve and golangci-lint to a
helper script to fix a problem with permissions on the /go/pkg
directory. Running the same commands within the Dockerfile would cause
the /go/pkg directory to be owned by root, preventing Go modules from
being installed by the `vscode` user.